requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --index-url=https://mirrors.aliyun.com/pypi/simple/
#
--index-url https://mirrors.aliyun.com/pypi/simple/

asgiref==3.3.1
    # via
    #   channels
    #   daphne
    #   django
attrs==20.3.0
    # via
    #   automat
    #   service-identity
    #   twisted
autobahn==20.12.3
    # via daphne
automat==20.2.0
    # via twisted
cffi==1.14.4
    # via cryptography
channels==3.0.3
    # via -r requirements.in
click==7.1.2
    # via pip-tools
constantly==15.1.0
    # via twisted
cryptography==3.3.1
    # via
    #   autobahn
    #   pyopenssl
    #   service-identity
daphne==3.0.1
    # via channels
django==3.1.5
    # via
    #   -r requirements.in
    #   channels
    #   djangorestframework
djangorestframework-jwt==1.11.0
    # via -r requirements.in
djangorestframework==3.12.2
    # via -r requirements.in
hyperlink==21.0.0
    # via
    #   autobahn
    #   twisted
idna==3.1
    # via
    #   hyperlink
    #   twisted
incremental==17.5.0
    # via twisted
pip-tools==5.5.0
    # via -r requirements.in
pyasn1-modules==0.2.8
    # via service-identity
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   service-identity
pycparser==2.20
    # via cffi
pyhamcrest==2.0.2
    # via twisted
pyjwt==1.7.1
    # via djangorestframework-jwt
pyopenssl==20.0.1
    # via twisted
pytz==2020.5
    # via django
service-identity==18.1.0
    # via twisted
six==1.15.0
    # via
    #   automat
    #   cryptography
    #   pyopenssl
sqlparse==0.4.1
    # via django
twisted[tls]==20.3.0
    # via daphne
txaio==20.12.1
    # via autobahn
zope.interface==5.2.0
    # via twisted

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
